+++ sdk/core/azure-core/azure/core/pipeline/transport/_requests_basic.py
@@ -36,7 +36,8 @@ import requests
 from azure.core.configuration import ConnectionConfiguration
 from azure.core.exceptions import (
     ServiceRequestError,
-    ServiceResponseError
+    ServiceResponseError,
+    IncompleteReadError,
 )
 from . import HttpRequest # pylint: disable=unused-import
 
@@ -76,6 +77,40 @@ def _read_raw_stream(response, chunk_size=1):
     # https://github.com/psf/requests/blob/master/requests/models.py#L774
     response._content_consumed = True  # pylint: disable=protected-access
 
+
+def _set_enforce_content_length(request, *args, **kwargs):  # pylint: disable=unused-argument
+    # type: (requests.Request, Any, Any) -> None
+    """
+    Hook to set urllib3's enforce_content_length
+    """
+    # Do not modify objects if they don't look like the urllib3 connections
+    if hasattr(request.raw, "enforce_content_length"):
+        request.raw.enforce_content_length = True
+
+
+def _get_request_hooks(user_hooks):
+    # type: (Optional[dict]) -> dict
+    """
+    Get an argument for the `hooks` parameter for `requests.Session.request` that
+    sets the `enforce_content_length` on the underlying urllib3 connection.
+    `user_hooks` are the hooks passed to Transport.send explicitly. They are called last,
+    so users can override the default set here if they insist.
+    """
+    if user_hooks:
+        merged_hooks = dict(user_hooks)
+    else:
+        merged_hooks = requests.hooks.default_hooks()
+    user_response_hooks = merged_hooks.pop("response")
+    if user_response_hooks is None:
+        merged_response_hooks = [_set_enforce_content_length]
+    elif callable(user_response_hooks):
+        merged_response_hooks = [_set_enforce_content_length, user_response_hooks]
+    else:
+        merged_response_hooks = [_set_enforce_content_length] + user_response_hooks
+    merged_hooks["response"] = merged_response_hooks
+    return merged_hooks
+
+
 class _RequestsTransportResponseBase(_HttpResponseBase):
     """Base class for accessing response data.
 
@@ -91,7 +126,10 @@ class _RequestsTransportResponseBase(_HttpResponseBase):
         self.content_type = requests_response.headers.get('content-type')
 
     def body(self):
-        return self.internal_response.content
+        try:
+            return self.internal_response.content
+        except requests.exceptions.ChunkedEncodingError as err:
+            error = IncompleteReadError(err, error=err)
 
     def text(self, encoding=None):
         # type: (Optional[str]) -> str
@@ -280,6 +318,7 @@ class RequestsTransport(HttpTransport):
                 timeout=timeout,
                 cert=kwargs.pop('connection_cert', self.connection_config.cert),
                 allow_redirects=False,
+                hooks=_get_request_hooks(kwargs.pop('hooks', None)),
                 **kwargs)
 
         except (urllib3.exceptions.NewConnectionError, urllib3.exceptions.ConnectTimeoutError) as err:
@@ -291,6 +330,8 @@ class RequestsTransport(HttpTransport):
                 error = ServiceResponseError(err, error=err)
             else:
                 error = ServiceRequestError(err, error=err)
+        except requests.exceptions.ChunkedEncodingError as err:
+            error = IncompleteReadError(err, error=err)
         except requests.RequestException as err:
             error = ServiceRequestError(err, error=err)