+++ sdk/core/azure-core/azure/core/pipeline/transport/_requests_trio.py
@@ -34,7 +34,8 @@ import requests
 
 from azure.core.exceptions import (
     ServiceRequestError,
-    ServiceResponseError
+    ServiceResponseError,
+    IncompleteReadError,
 )
 from azure.core.pipeline import Pipeline
 from ._base import HttpRequest
@@ -42,7 +43,7 @@ from ._base_async import (
     AsyncHttpResponse,
     _ResponseStopIteration,
     _iterate_response_content)
-from ._requests_basic import RequestsTransportResponse, _read_raw_stream
+from ._requests_basic import RequestsTransportResponse, _read_raw_stream, _get_request_hooks
 from ._base_requests_async import RequestsAsyncTransportBase
 from .._tools import get_block_size as _get_block_size, get_internal_response as _get_internal_response
 
@@ -164,6 +165,7 @@ class TrioRequestsTransport(RequestsAsyncTransportBase):  # type: ignore
                         timeout=kwargs.pop('connection_timeout', self.connection_config.timeout),
                         cert=kwargs.pop('connection_cert', self.connection_config.cert),
                         allow_redirects=False,
+                        hooks=_get_request_hooks(kwargs.pop('hooks', None)),
                         **kwargs),
                     limiter=trio_limiter)
             except AttributeError:  # trio < 0.12.1
@@ -179,6 +181,7 @@ class TrioRequestsTransport(RequestsAsyncTransportBase):  # type: ignore
                         timeout=kwargs.pop('connection_timeout', self.connection_config.timeout),
                         cert=kwargs.pop('connection_cert', self.connection_config.cert),
                         allow_redirects=False,
+                        hooks=_get_request_hooks(kwargs.pop('hooks', None)),
                         **kwargs),
                     limiter=trio_limiter)
 
@@ -191,6 +194,8 @@ class TrioRequestsTransport(RequestsAsyncTransportBase):  # type: ignore
                 error = ServiceResponseError(err, error=err)
             else:
                 error = ServiceRequestError(err, error=err)
+        except requests.exceptions.ChunkedEncodingError as err:
+            error = IncompleteReadError(err, error=err)
         except requests.RequestException as err:
             error = ServiceRequestError(err, error=err)